changed many "unknown error" status messages to support the new unknown error dialog added exclamation mark to apartment gateway count if 0 restructured isValid() method to prevent errors all apartments can now be deleted properly

+++ Shared/src/main/java/eu/power_switch/shared/constants/SettingsConstants.java
@@ -45,7 +45,7 @@ public class SettingsConstants {
 
     public static final int DEFAULT_VIBRATION_DURATION_HAPTIC_FEEDBACK = 75;
 
-    public static final int INVALID_APARTMENT_ID = -1;
+    public static final long INVALID_APARTMENT_ID = -1;
 
     // Main Tabs
     public static final int ROOMS_TAB_INDEX = 0;
